cast image to float32

+++ torch/torch-imfit5.py
@@ -343,7 +343,7 @@ class ImageFitter:
     
     def sobel_filter(self, image):
         # Ensure image is in the right format (B, C, H, W)
-        image = image.unsqueeze(0)
+        image = image.unsqueeze(0).float()
     # Define base Sobel kernels
         sobel_x = torch.tensor([[-1, 0, 1],
                             [-2, 0, 2],
